requirements.txt
pytz==2018.5  # https://github.com/stub42/pytz

# Django
# ------------------------------------------------------------------------------
# django==2.0.8  # pyup: < 2.1  # https://www.djangoproject.com/
Django==2.1.5
mysqlclient==1.3.14
django-environ==0.4.5  # https://github.com/joke2k/django-environ

# Django REST Framework
djangorestframework-jwt==1.11.0
djangorestframework==3.8.2  # https://github.com/encode/django-rest-framework
django-rest-swagger # https://django-rest-swagger.readthedocs.io/en/latest/
drf-yasg==1.2.1
django-url-filter==0.3.12
django-cors-headers==2.5.3
Pygments==2.4.0

# AWS
django-storages==1.7.1
boto3==1.9.65
botocore==1.12.65

# Image processor
django-imagekit==4.0.2
pilkit==2.0
Pillow==5.3.0

# Celery for job queue
celery==4.2.1

# Social login
social-auth-core>=3.0,<4.0
social-auth-app-django>=3.1,<4.0
rest-social-auth>=2.1.0,<3.0
